requirements.txt
datasets
bitsandbytes
torch
transformers
trl
accelerate
evaluate
jiwer
Levenshtein
kornia
pillow
numpy
tqdm
safetensors
scipy
